requirements.txt
matplotlib==3.4.2
numpy==1.18.5
pandas==1.2.4
Pillow==8.4.0
scikit_learn==1.0.1
seaborn==0.11.1
sklearn_genetic==0.4.1
streamlit==1.2.0
